Add a LongIterator interface to favor primitives.
The current IteratorLong class is a nested class in
NonBlockingHashMapLong, which means to iterate over the keys without
auto boxing the casting can get awkward. This adds an interface
LongIterator which the returned iterator from keySet().iterator() can be
cast to safely. This also cleans up the Maven POM file and adds a simple
test for the iterator.